+++ packages/google-cloud-parallelstore/google/cloud/parallelstore_v1beta/services/parallelstore/transports/grpc_asyncio.py
@@ -286,7 +286,7 @@ class ParallelstoreGrpcAsyncIOTransport(ParallelstoreTransport):
     ]:
         r"""Return a callable for the list instances method over gRPC.
 
-        Lists Instances in a given project and location.
+        Lists all instances in a given project and location.
 
         Returns:
             Callable[[~.ListInstancesRequest],
@@ -314,7 +314,7 @@ class ParallelstoreGrpcAsyncIOTransport(ParallelstoreTransport):
     ]:
         r"""Return a callable for the get instance method over gRPC.
 
-        Gets details of a single Instance.
+        Gets details of a single instance.
 
         Returns:
             Callable[[~.GetInstanceRequest],
@@ -371,7 +371,7 @@ class ParallelstoreGrpcAsyncIOTransport(ParallelstoreTransport):
     ]:
         r"""Return a callable for the update instance method over gRPC.
 
-        Updates the parameters of a single Instance.
+        Updates the parameters of a single instance.
 
         Returns:
             Callable[[~.UpdateInstanceRequest],
@@ -399,7 +399,7 @@ class ParallelstoreGrpcAsyncIOTransport(ParallelstoreTransport):
     ]:
         r"""Return a callable for the delete instance method over gRPC.
 
-        Deletes a single Instance.
+        Deletes a single instance.
 
         Returns:
             Callable[[~.DeleteInstanceRequest],
@@ -427,8 +427,7 @@ class ParallelstoreGrpcAsyncIOTransport(ParallelstoreTransport):
     ]:
         r"""Return a callable for the import data method over gRPC.
 
-        ImportData copies data from Cloud Storage to
-        Parallelstore.
+        Copies data from Cloud Storage to Parallelstore.
 
         Returns:
             Callable[[~.ImportDataRequest],
@@ -456,8 +455,7 @@ class ParallelstoreGrpcAsyncIOTransport(ParallelstoreTransport):
     ]:
         r"""Return a callable for the export data method over gRPC.
 
-        ExportData copies data from Parallelstore to Cloud
-        Storage
+        Copies data from Parallelstore to Cloud Storage.
 
         Returns:
             Callable[[~.ExportDataRequest],