+++ packages/google-shopping-merchant-accounts/google/shopping/merchant_accounts_v1beta/services/accounts_service/transports/grpc.py
@@ -297,9 +297,12 @@ class AccountsServiceGrpcTransport(AccountsServiceTransport):
         r"""Return a callable for the delete account method over gRPC.
 
         Deletes the specified account regardless of its type:
-        standalone, MCA or sub-account. Deleting an MCA leads to
-        the deletion of all of its sub-accounts. Executing this
-        method requires admin access.
+        standalone, MCA or sub-account. Deleting an MCA leads to the
+        deletion of all of its sub-accounts. Executing this method
+        requires admin access. The deletion succeeds only if the account
+        does not provide services to any other account and has no
+        processed offers. You can use the ``force`` parameter to
+        override this.
 
         Returns:
             Callable[[~.DeleteAccountRequest],
@@ -358,7 +361,9 @@ class AccountsServiceGrpcTransport(AccountsServiceTransport):
         size or filters. This is not just listing the
         sub-accounts of an MCA, but all accounts the calling
         user has access to including other MCAs, linked
-        accounts, standalone accounts and so on.
+        accounts, standalone accounts and so on. If no filter is
+        provided, then it returns accounts the user is directly
+        added to.
 
         Returns:
             Callable[[~.ListAccountsRequest],